requirements.txt
amqp==5.0.6
appdirs==1.4.4
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.4.0
black==21.6b0
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dateparser==1.0.0
decorator==5.0.9
Django==3.2.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==2.10
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.1.0
matplotlib-inline==0.1.2
mypy-extensions==0.4.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
regex==2021.7.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
